requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.0
pandas>=2.0.0
Pillow>=10.0.0
watchdog>=3.0.0
python-dotenv
opencv-python-headless